Expand HingeJoint and Cylinder to use ContinuousMap
Instead of using a single float value to define the radius of a
cylinder, use a ContinuousMap<float, float> which allows for a
varying radius depending on the position along the length of the
cylinder. Also adapt HingeJoint and Capsule to make use of these
changes. This way, any derived type of function that inherits from
ContinuousMap<float, float> can be used to describe the radius of
the cylinders.