Set up Godot project and rendering bugfixes
Set up Godot project for rendering, fixing some render bugs in the
engine.

The files within the "gui/" directory are examples on how to use
the engine to render output, and are subject to change.

Within the "engine/" directory, a few changes are made to enable
a frontend or render engine to better output a mesh.

Also formatting changes are made, enforcing the use of tabs for
indentation throughout the project.