Add Get functions for the position and normal of a Curve
Add GetPositionAt(float t) and GetNormalAt(float t) for all Curve
classes, which are required for raytracing eventually. (The rays
will be cast from the GetPositionAt() position along a curve in the
direction of GetNormalAt(float t).

Make a Curve an implementation of ContinuousMap<float, Vector3> for
future use. Depending on refactoring this might or might not be
useful, to be reviewed later.