Add QuadraticFunction class
Add a QuadraticFunction class that describes a quadratic function,
which is useful to model organically flowing curves and surfaces.
A special property of a quadratic function over a cubic function is
that raytracing is still analytically solvable, while for cubic
functions one would have to resort to numerical methods that will
slow down the program.