Change Capsule to take center line and height map instead of Cylinder instance
The Capsule constructor used to take a full Cylinder instance to define its
shaft shape. However, with the new 2D function defining the radius of a shaft,
it is now impossible to get a singular value for the end cap radius from the
shaft radius at the end points, since the height map value can now be
different for different angles at the ends of a Cylinder. Defining the center
curve and height map separately would resolve this issue.

Current commit does not fully implement a 2D height map for the end caps yet,
but that will be added in a later commit.