Shatter project into a zillion pieces and rearrange so Godot works.
Each subcomponent of Engine is its own project now. I'm not sure that's the structure we want to go with, but it allows us to put an Engine.sln of its own in /engine/ without recreating the exact same problem. But I'm not sure we want to. Hm, thinking on this more...