ComposedMap: f ∘ g for ContinuousMaps, basically
Also adds ContinuousMapExtensions.cs, providing fake member functions for ContinuousMap (for functions that don't need any inside details on the maps). Note that extensions will continue to work if we refactor ContinuousMap to IContinousMap, whereas member functions with bodies do not.

The constructor for ComposedMap is ugly because new ComposedMap<in, intermediate, out> is long to type, but automatic type deduction for method calls means the convenience methods in ContinuousMapExtensions.cs do not need explicit type parameters, so use them instead.