Add tests for .Equals and ==.
Assert.Equal relies on IEquatable<T>, which uses the Equals method, which is why it's possible to expect something to be equal to NaN, which isn't equal to itself using the `==` arithmetic operator.

For the tests for addition to be meaningful, the assertion itself has to be meaningful - which means the .Equals function has to be meaningful. So I've added a test for that, and while I'm at it, another one that puts the == operator through its paces, too, with the same cases - including highlighting the different behavior for NaN vs. NaN.