Allow shifted and rescaled SymmetricCylinder to be raytraced
Rescale and shift the rays used for raytracing a shifted or
scaled SymmetricCylinder so that it correctly raytraces it.

This can be implemented using an inverse matrix multiplication
as well, but for now this seems like the fastest option.

Also fix small issue where the wrong coordinates were taken from
the end cap of the Capsule surface.