Convenience tools for converting functions to ContinuousMaps.
I couldn't find a way to make the syntax lower-friction than this, everything requires explicitly specifying <TIn, TOut> somewhere. I think I could make one that doesn't require writing out the generic types again as a static function of a class that is itself not generic, but it's not obvious that there's any reasonable name for such a class that is actually less annoying than the average (Func<TIn, TOut>) cast.