Add failing test case where the function does not handle "concordant infinity" correctly.
In IEEE floating-point math, indeterminate forms are represented as NaN. That includes all `0 * infinity` expressions. IEEE float does not "trust its zero" to be a true zero; it represents that zero can be an underflow, making IEEE zero behave more like a "limit as function approaches zero here" rather than a true zero, reflecting its lack of confidence that zero is zero.

That makes sense at the scope of an individual multiplication operation. However, in the context of an entire quadratic expression, we might know more. `(positive zero) * (signed infinity)` is, mathematically, an unknown value between true zero and infinity of the same sign as provided. (For IEEE negative zero, the sign flips.) If a term is IEEE NaN because it is an unknown and possible infinite value _of known sign_, adding it to an infinity of the _same sign_ "should", mathematically, produce infinity of the same sign. However, once this value has been represented as NaN, our mathematical knowledge about the (potentially transfinite) values it might represent has been lost, causing us to get NaN results in cases where, mathematically, we should get a definite infinity instead.

Possible solutions:

* Special-case handling of all `infinity * 0` cases. (I've had a class kicking around in the back of my mind for these.)
* Switch from `float` to `Real` and, in Real, implement a _much_ more comprehensive model of indefinite values, which would accurately represent "a nonnegative number of completely unknown magnitude" and correctly reconcile it with infinities of various signs. (I've had _that_ kicking around in my mind as a hobby project.)
* Throw ArgumentException on transfinite (NaN or infinite) coefficients or values for `x` and declare the entire thing to be out of range. This requires upgrading overflows to an OverflowException rather than using IEEE representations for out-of-range values.
* Convert all transfinite results to NaN.

The current status quo is that some expressions will correctly give infinity (depending on whether any coefficients are 0, if X is one of the infinite values) and others will give NaN instead. Inconsistency is bad; either be consistently correct, consistently NaN, or reject the input.