Discard rays outside of the pie slice of the SymmetricCylinder
A SymmetricCylinder has a StartAngle and EndAngle, which defines a
pie-slice of the cylinder that can be raytraced. These start and
end angles basically define the maximal hinge angles in the case
of a hinge joint. The raytracer didn't yet account for that,
but now it ensures that only the pie-slice is raytraced that is required.