Make sure raytracer can handle rotated objects
The raytracer of the SymmetricCylinder (which is used for HingeJoint objects)
assumed that the SymmetricCylinder was always at length 1, positioned
horizontally in the z-direction. However, a SymmetricCylinder may take
any orientation, and the raytracer should be able to handle all such
cases.

By transforming the rays such that the mathematical representation
of the SymmetricCylinder is scaled, rotated, and translated properly,
the raytracer can now handle SymmetricCylinders that are positioned
and oriented in any arbitrary direction.