Fix solution file for mono & MonoDevelop (#53)
* Fix .sln configuration for MonoDevelop

Also ensures that UnitTests are not built during the ExportRelease
and ExportDebug builds.

* Fix compiler complaining about #ifdef after first token

This might be a difference between mono and .NET Core. In mono, the
compiler complains that #ifdef cannot be called after the first
token. In the case of RealTests.cs, the tokens 'using System;' and such
were called before the preprocessor statements. Swapping them fixes the
issue.