Add index-centric behaviors.
Since MutableCurve (unlike FlexSpline) doesn't sort-in-place, indexes of points are stable unless a point is removed. Therefore, it is reasonable to offer some list-like features by providing a Count, an indexer, and indexed removal. We're still not going to allow assignment into the list because mutable points need appropriate callback.