Replace Real with double.
Real was an interesting idea but not what we need, and it prevents us from using other peoples' vector math libraries.

Note that two of the TestValueCalculation tests in QuadraticFunctionTests now fail - I think they might've been failing before? in any case, it's an edge condtion that might be nice to get right, but everything else seems fine after this change - as far as the unit tests can tell, at least.